requirements.txt
annotated-types==0.6.0
anyio==4.2.0
build==1.0.3
CacheControl==0.13.1
certifi==2023.11.17
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.8
dulwich==0.21.7
fastapi==0.109.0
fastjsonschema==2.19.1
filelock==3.13.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
hiredis==2.3.2
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
keyring==24.3.0
more-itertools==10.2.0
msgpack==1.0.7
mwoauth==0.4.0
oauthlib==3.2.2
packaging==23.2
pexpect==4.9.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
ptyprocess==0.7.0
pydantic==2.6.0
pydantic_core==2.16.1
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dotenv==1.0.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==3.6.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
shellingham==1.5.4
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.35.1
tomlkit==0.12.3
trove-classifiers==2024.1.8
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.27.0.post1
virtualenv==20.25.0
watchfiles==0.21.0
websockets==12.0
